+++ src/org/nschmidt/abalone/PlayerLogic.java
@@ -1,6 +1,8 @@
 package org.nschmidt.abalone;
 
+import java.math.BigInteger;
+
 public interface PlayerLogic {
     public Player switchPlayer();
-    public long getNumber();
+    public BigInteger getNumber();
 }
