+++ src/org/nschmidt/abalone/test/TripleMoverTest.java
@@ -4,21 +4,23 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.nschmidt.abalone.Player.BLACK;
 import static org.nschmidt.abalone.TripleMover.moveThreeMarbles;
 
+import java.math.BigInteger;
+
 import org.junit.jupiter.api.Test;
 
 class TripleMoverTest {
 
     @Test
     void testMoveThreeMarbles1() {
-        final long state = 104340773383259005L;
-        long[] moves = moveThreeMarbles(state, BLACK);
+        final BigInteger state = BigInteger.valueOf(104340773383259005L);
+        BigInteger[] moves = moveThreeMarbles(state, BLACK);
         assertEquals(6, moves.length);
     }
     
     @Test
     void testMoveThreeMarbles2() {
-        final long state = 1804882274845767L;
-        long[] moves = moveThreeMarbles(state, BLACK);
+        final BigInteger state = BigInteger.valueOf(1804882274845767L);
+        BigInteger[] moves = moveThreeMarbles(state, BLACK);
         assertEquals(6, moves.length);
     }
 }
